Introducing operators module
