+++ spec/ExpectationsCorrectorSpec.hs
@@ -1,69 +1,41 @@
-{-# LANGUAGE TupleSections #-}
-
 module ExpectationsCorrectorSpec (spec) where
 
 import           Test.Hspec
 import           Language.Mulang.Analyzer hiding (spec)
+import           Language.Mulang.Operators
+import           Language.Mulang.Operators.Haskell (haskellTokensTable)
+import           Language.Mulang.Operators.Ruby (rubyTokensTable)
+import           Language.Mulang.Operators.Java (javaTokensTable)
+import           Language.Mulang.Operators.Python (pythonTokensTable)
 import           Language.Mulang.Ast (PrimitiveOperator (..))
-import           Control.Applicative ((<|>))
 
-import           Data.Tuple (swap)
+import           Control.Applicative ((<|>))
+import           Control.Monad ((>=>))
+
+import           Text.Read (readMaybe)
+import           Data.List (stripPrefix)
 import           Data.Map.Strict (Map)
 import qualified Data.Map.Strict as Map
 
 type Inspection = String
-type Token = String
-
-type TokensTable = Map PrimitiveOperator [Token]
-type OperatorsTable = Map Token PrimitiveOperator
 type KeywordInspectionsTable = Map Token Inspection
 
--- C-style tokens
-defaultTokensTable :: TokensTable
-defaultTokensTable = Map.fromList [
-  (Equal, ["=="]),
-  (NotEqual, ["!="]),
-  (Negation, ["!"]),
-  (And, ["&&"]),
-  (Or, ["||"]),
-  (GreatherOrEqualThan, [">="]),
-  (GreatherThan, [">"]),
-  (LessOrEqualThan, ["<="]),
-  (LessThan, ["<"])
- ]
+buildUsageInspection :: PrimitiveOperator -> Inspection
+buildUsageInspection = ("Uses" ++) . show
 
-buildTokensTable :: [(PrimitiveOperator, [Token])] -> TokensTable
-buildTokensTable = flip Map.union defaultTokensTable  . Map.fromList
+unbuildUsageInspection :: Inspection -> Maybe PrimitiveOperator
+unbuildUsageInspection = stripPrefix "Uses" >=> readMaybe
+
+unbuildDeclarationInspection :: Inspection -> Maybe PrimitiveOperator
+unbuildDeclarationInspection = stripPrefix "Declares" >=> readMaybe
+
+primitiveUsage = unbuildUsageInspection
 
 tokensTable :: Language -> TokensTable
-tokensTable Haskell = buildTokensTable [
-    (NotEqual, ["/="]),
-    (Negation, ["not"]),
-    (Otherwise, ["otherwise"]),
-    (BackwardComposition, ["."])
-  ]
-tokensTable Java = buildTokensTable [
-    (Hash, ["hashCode"])
-  ]
-tokensTable Ruby = buildTokensTable [
-    (And, ["&&", "and"]),
-    (Or, ["||", "or"]),
-    (Hash, ["hash"]),
-    (ForwardComposition, [">>"]),
-    (BackwardComposition, ["<<"])
-  ]
-tokensTable Python = buildTokensTable [
-    (NotEqual, ["!=", "<>"]),
-    (Negation, ["not"]),
-    (And, ["and"]),
-    (Or, ["or"]),
-    (Hash, ["hash"])
-  ]
-
-operatorsTable :: Language -> OperatorsTable
-operatorsTable =  Map.fromList . concatMap (fill . swap) . Map.toList . tokensTable
-  where
-    fill (xs, t) = map (,t) xs
+tokensTable Haskell = haskellTokensTable
+tokensTable Java = javaTokensTable
+tokensTable Ruby = rubyTokensTable
+tokensTable Python = pythonTokensTable
 
 keywordInspectionsTable :: Language -> KeywordInspectionsTable
 keywordInspectionsTable Haskell = Map.fromList [
@@ -90,14 +62,14 @@ keywordInspectionsTable Python = Map.fromList [
   ("for", "UsesForeach")
  ]
 
-lookupOperatorInspection :: Language -> Token -> Maybe Inspection
-lookupOperatorInspection language target = fmap (("Uses" ++) . show) . (Map.lookup target) . operatorsTable $ language
+operatorInspection :: Language -> Token -> Maybe Inspection
+operatorInspection language target = fmap buildUsageInspection . (Map.lookup target) . buildOperatorsTable . tokensTable $ language
 
-lookupKeywordInspection :: Language -> Token -> Maybe Inspection
-lookupKeywordInspection language target = Map.lookup target . keywordInspectionsTable $ language
+keywordInspection :: Language -> Token -> Maybe Inspection
+keywordInspection language target = Map.lookup target . keywordInspectionsTable $ language
 
 run :: Language -> Token -> Maybe Inspection
-run language  target = lookupOperatorInspection language target <|> lookupKeywordInspection language target
+run language  target = operatorInspection language target <|> keywordInspection language target
 
 spec :: Spec
 spec = do
