+++ src/Language/Mulang/Operators/Haskell.hs
@@ -1,0 +1,14 @@
+{-# LANGUAGE ViewPatterns #-}
+
+module Language.Mulang.Operators.Haskell (haskellTokensTable) where
+import Language.Mulang.Operators (TokensTable, buildTokensTable)
+import Language.Mulang.Ast (PrimitiveOperator (..))
+
+haskellTokensTable :: TokensTable
+haskellTokensTable =
+  buildTokensTable [
+    (NotEqual, ["/="]),
+    (Negation, ["not"]),
+    (Otherwise, ["otherwise"]),
+    (BackwardComposition, ["."])
+  ]
