+++ src/Language/Mulang/Operators/Python.hs
@@ -1,0 +1,15 @@
+{-# LANGUAGE ViewPatterns #-}
+
+module Language.Mulang.Operators.Python (pythonTokensTable) where
+import Language.Mulang.Operators (TokensTable, buildTokensTable)
+import Language.Mulang.Ast (PrimitiveOperator (..))
+
+pythonTokensTable :: TokensTable
+pythonTokensTable =
+  buildTokensTable [
+    (NotEqual, ["!=", "<>"]),
+    (Negation, ["not"]),
+    (And, ["and"]),
+    (Or, ["or"]),
+    (Hash, ["hash"])
+  ]
